+++ src/components/Menus/Sidebar/InstructionsPanel/InstructionsPanel.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from "react";
+import React, { useEffect, useRef, useMemo } from "react";
 import SidebarPanel from "../SidebarPanel";
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
@@ -10,12 +10,18 @@ import "prismjs/plugins/line-numbers/prism-line-numbers.css";
 import "prismjs/plugins/line-highlight/prism-line-highlight.css";
 
 const InstructionsPanel = () => {
-  const steps = useSelector((state) => state.instructions.project.steps);
+  const steps = useSelector((state) => state.instructions.project?.steps);
+  const quiz = useSelector((state) => state.instructions?.quiz);
+
   const currentStepPosition = useSelector(
     (state) => state.instructions.currentStepPosition,
   );
   const { t } = useTranslation();
   const stepContent = useRef();
+
+  const isQuiz = useMemo(() => {
+    return !!quiz?.questionCount;
+  }, [quiz]);
 
   const applySyntaxHighlighting = (container) => {
     const codeElements = container.querySelectorAll(".language-python");
@@ -26,17 +32,23 @@ const InstructionsPanel = () => {
   };
 
   useEffect(() => {
-    if (steps[currentStepPosition]) {
+    const setStepContent = (content) => {
       stepContent.current.parentElement.scrollTo({ top: 0 });
-      stepContent.current.innerHTML = steps[currentStepPosition].content;
+      stepContent.current.innerHTML = content;
       applySyntaxHighlighting(stepContent.current);
+    };
+
+    if (isQuiz) {
+      setStepContent(quiz.questions[quiz.currentQuestion]);
+    } else if (steps[currentStepPosition]) {
+      setStepContent(steps[currentStepPosition].content);
     }
-  }, [steps, currentStepPosition]);
+  }, [steps, currentStepPosition, quiz, isQuiz]);
 
   return (
     <SidebarPanel
       heading={t("instructionsPanel.projectSteps")}
-      Footer={ProgressBar}
+      Footer={!isQuiz && ProgressBar}
     >
       <div className="project-instructions" ref={stepContent}></div>
     </SidebarPanel>